+++ R/guidelines.R
@@ -53,17 +53,3 @@ guidelines <- function(
   data <- add_class(data, "dynguidelines::guidelines")
   data
 }
-
-#' Check whether object is guidelines
-#'
-#' @param guidelines The object to check
-#' @export
-is_guidelines <- function(guidelines) {
-  if("dynguidelines::guidelines" %in% class(guidelines)) {
-    TRUE
-  } else if (all(c("methods", "answers") %in% names(guidelines))) {
-    TRUE
-  } else {
-    FALSE
-  }
-}