+++ inst/NEWS.md
@@ -11,6 +11,7 @@
 * Fix javascript error for lenses which activated wrong columns
 * Fix bug when no n_cells or n_features is entered
 * Clean stability column names
+* Clean scalability column names
 
 # dynguidelines 0.3.0 (15-11-2018)