+++ inst/NEWS.md
@@ -13,6 +13,7 @@
 * Clean stability column names
 * Clean scalability column names
 * Add tooltips to priors
+* Small fixes for default columns
 
 # dynguidelines 0.3.0 (15-11-2018)